Header component is used as common component for all other components.
