+++ src/components/Grocery.jsx
@@ -1,12 +1,8 @@
 import React from "react";
-import Header from "./Header";
 
 const Grocery = () => {
   return (
-    <div>
-      <Header />
-      Gcorery component this compannet has so many sub components.
-    </div>
+    <div>Gcorery component this compannet has so many sub components.</div>
   );
 };
