+++ src/components/HeaderSubRoutes/About.jsx
@@ -1,10 +1,8 @@
 import React from "react";
-import Header from "../Header";
 
 const About = () => {
   return (
     <div>
-      <Header />
       <div className="body-sec">About Page</div>
     </div>
   );
