+++ src/components/HeaderSubRoutes/Cart.jsx
@@ -1,10 +1,8 @@
 import React from "react";
-import Header from "../Header";
 
 const Cart = () => {
   return (
     <div>
-      <Header />
       <div className="body-sec">Cart Page</div>
     </div>
   );
