+++ src/components/HeaderSubRoutes/Contact.jsx
@@ -1,10 +1,8 @@
 import React from "react";
-import Header from "../Header";
 
 const Contact = () => {
   return (
     <div>
-      <Header />
       <div className="body-sec">Contact Page</div>
     </div>
   );
